fix: fix diagnostics test and some minor refactoring
