+++ analyzers/diagnostics/diagnostics.go
@@ -2,18 +2,13 @@ package diagnostics
 
 import (
 	"fmt"
-	"os"
 	"path/filepath"
-	"regexp"
 	"sort"
 	"strings"
 
 	"github.com/deepsourcelabs/cli/analyzers/validator"
 	"github.com/morikuni/aec"
 )
-
-// fieldRegexp is a regular expression for checking the current line is a field or not.
-var fieldRegexp = `\s*(?P<field>.+)\s*=\s*`
 
 // FileDiagnostic represents the diagnostics present in a file.
 type FileDiagnostic struct {
@@ -69,10 +64,8 @@ func groupDiagnostics(diagnostics []Diagnostic) []FileDiagnostic {
 			Filename:    filename,
 			Diagnostics: groupedDiagnosticsMap[filename],
 		}
-
 		groupedDiagnostics = append(groupedDiagnostics, fileDiagnostic)
 	}
-
 	return groupedDiagnostics
 }
 
@@ -131,16 +124,6 @@ func constructDiagnostic(diag Diagnostic) string {
 	return diagnosticMsg
 }
 
-// readFileContent reads the file and returns its content.
-func readFileContent(filename string) (string, error) {
-	content, err := os.ReadFile(filename)
-	if err != nil {
-		return "", err
-	}
-
-	return string(content), nil
-}
-
 // getDiagnosticsFromFile uses the file content to return diagnostics with metadata like line number, content, etc.
 func getDiagnosticsFromFile(filename string, errors []validator.ErrorMeta) []Diagnostic {
 	fileContent, err := readFileContent(filename)
@@ -150,15 +133,16 @@ func getDiagnosticsFromFile(filename string, errors []validator.ErrorMeta) []Dia
 
 	diagnostics := []Diagnostic{}
 
-	// Get filename (base path) from absolute file path.
+	// Get filename (base path) from absolute file path and get its lines data post trimming any
+	// trailing spaces.
 	filename = filepath.Base(filename)
-
-	lines := strings.Split(string(fileContent), "\n")
+	lines := strings.Split(strings.TrimRight(fileContent, " "), "\n")
 
 	// Iterate over each error and check line-by-line.
 	for _, err := range errors {
 		for lineNum, line := range lines {
-			// If the field is empty, do NOT render the codeframe.
+
+			// Case 1: If the field is empty, do NOT render the codeframe.
 			if err.Field == "" {
 				diag := Diagnostic{
 					Filename:     filename,
@@ -166,13 +150,13 @@ func getDiagnosticsFromFile(filename string, errors []validator.ErrorMeta) []Dia
 					Codeframe:    "",
 					ErrorMessage: err.Message,
 				}
-
 				diagnostics = append(diagnostics, diag)
 				break
 			}
 
-			containsField := checkField(line, err.Field)
-			if containsField {
+			// Case 2: Check if the invalid field pointed by the error is present
+			// in the current line. If yes, prepare its codeframe.
+			if checkField(line, err.Field) {
 				// Prepare code frame for the current line.
 				codeFrame := prepareCodeFrame(lineNum, lines)
 
@@ -182,60 +166,9 @@ func getDiagnosticsFromFile(filename string, errors []validator.ErrorMeta) []Dia
 					Codeframe:    codeFrame,
 					ErrorMessage: err.Message,
 				}
-
 				diagnostics = append(diagnostics, diag)
 			}
 		}
 	}
-
 	return diagnostics
-}
-
-// prepareCodeFrame prepares a code frame using the file content. The code frame is meant to be displayed on the console while reporting diagnostics.
-// NOTE: lineNum always starts from 0.
-func prepareCodeFrame(lineNum int, lines []string) string {
-	frame := ""
-
-	if lineNum <= 1 {
-		// Case 1: When the line is near the top of the file.
-		// Generate a frame with the current and next line only.
-		frame += aec.LightRedF.Apply(fmt.Sprintf("> %d | %s\n", lineNum+1, lines[lineNum]))
-		frame += fmt.Sprintf("  %d | %s\n", lineNum+2, lines[lineNum+1])
-
-	} else if lineNum >= (len(lines) - 1) {
-		// Case 2: When the line is near the bottom of the file.
-		// Generate a frame with the current line only.
-		frame += aec.LightRedF.Apply(fmt.Sprintf("> %d | %s\n", lineNum, lines[lineNum-1]))
-	} else {
-		// Case 3: When the line is in between the top and the bottom.
-		// Generate a frame with the the previous, current and the next line.
-		frame += fmt.Sprintf("  %d | %s\n", lineNum, lines[lineNum-1])
-		frame += aec.LightRedF.Apply(fmt.Sprintf("> %d | %s\n", lineNum+1, lines[lineNum]))
-		frame += fmt.Sprintf("  %d | %s\n", lineNum+2, lines[lineNum+1])
-	}
-
-	return frame
-}
-
-// checkField checks if the line contains the field.
-func checkField(line, field string) bool {
-
-	// We use fieldRegexp for checking if the current line is a field or not.
-	// If it's not a field, for example, comments for explanation, etc., then checkField should return false.
-	exp := regexp.MustCompile(fieldRegexp)
-
-	// Get index for the "field" named group.
-	matches := exp.FindStringSubmatch(line)
-	groupIndex := exp.SubexpIndex("field")
-
-	// Capture match. We proceed only if the groupIndex isn't -1 (group not found), and if the number of matches aren't zero.
-	// Once done, we populate match with the match found using the regular expression.
-	match := ""
-	if len(matches) != 0 && groupIndex != -1 {
-		match = matches[groupIndex]
-	}
-	match = strings.TrimSpace(match)
-
-	// Return true if the match contains the field.
-	return strings.Contains(match, field)
 }
